+++ src/main/java/io/hhplus/conbook/application/concert/dto/ConcertBookingCommand.java
@@ -1,0 +1,18 @@
+package io.hhplus.conbook.application.concert.dto;
+
+import static io.hhplus.conbook.application.concert.DateValidator.validateDate;
+
+public class ConcertBookingCommand {
+    public record BookingSeat(
+            long concertId,
+            String date,
+            String userUUID,
+            long seatId
+    ) {
+        public BookingSeat {
+            validateDate(date);
+
+            if (seatId <= 0) throw new IllegalArgumentException();
+        }
+    }
+}
