+++ src/main/java/io/hhplus/conbook/application/concert/dto/ConcertBookingResult.java
@@ -1,0 +1,17 @@
+package io.hhplus.conbook.application.concert.dto;
+
+import lombok.Builder;
+
+import java.time.LocalDateTime;
+
+public class ConcertBookingResult {
+    @Builder
+    public record BookingSeat(
+            long bookingId,
+            String userName,
+            String rowName,
+            int seatNo,
+            LocalDateTime bookingDateTime,
+            LocalDateTime expirationTime
+    ) {}
+}
