+++ src/main/java/io/hhplus/conbook/application/concert/dto/ConcertCommand.java
@@ -1,0 +1,18 @@
+package io.hhplus.conbook.application.concert.dto;
+
+import static io.hhplus.conbook.application.concert.DateValidator.validateDate;
+
+public class ConcertCommand {
+    public record Search(
+            long concertId
+    ) {}
+
+    public record Status(
+            long concertId,
+            String date
+    ) {
+        public Status {
+            validateDate(date);
+        }
+    }
+}
