+++ src/test/java/io/hhplus/conbook/application/concert/ConcertBookingFacadeTest.java
@@ -1,5 +1,7 @@
 package io.hhplus.conbook.application.concert;
 
+import io.hhplus.conbook.application.concert.dto.ConcertBookingCommand;
+import io.hhplus.conbook.application.concert.dto.ConcertBookingResult;
 import io.hhplus.conbook.domain.user.User;
 import io.hhplus.conbook.domain.user.UserService;
 import org.junit.jupiter.api.DisplayName;
@@ -10,10 +12,10 @@ import org.springframework.boot.test.context.SpringBootTest;
 import static org.assertj.core.api.Assertions.assertThat;
 
 @SpringBootTest
-class ConcertFacadeTest {
+class ConcertBookingFacadeTest {
 
     @Autowired
-    ConcertFacade concertFacade;
+    ConcertBookingFacade concertBookingFacade;
     @Autowired
     UserService userService;
 
@@ -29,7 +31,7 @@ class ConcertFacadeTest {
         User userByUUID = userService.getUserByUUID(uuid);
 
         // when
-        ConcertResult.BookingSeat booking = concertFacade.bookConcertSeat(new ConcertCommand.Booking(concertId, date, uuid, seatId));
+        ConcertBookingResult.BookingSeat booking = concertBookingFacade.bookConcertSeat(new ConcertBookingCommand.BookingSeat(concertId, date, uuid, seatId));
 
         // then
         assertThat(booking.userName()).isEqualTo(userByUUID.getName());
